pyproject.toml
[tool.poetry]
name = "hltv-bot"
version = "0.2.0"
description = ""
authors = ["kiobu <kiobu@sdf.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
feedparser = "^6.0.10"
discord-py = "^2.3.2"
discord = "^2.3.2"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
